requirements.txt
matplotlib
requests
torch
fair-esm
biopython
joblib
compress_pickle
sklearn
pandas